Moved `format_capacity_title`
